Implement isGameOver, Add screen wipe, remove non-repeated output
Implements isGameOver, which checks the score, and outputs a congratulations and ends the run loop if it hits the target. Also adds a screen wipe for each loop. This allows a consistent input screen, but trades off not being able to output one-off messages easily, so those have been removed.